work around for flinging behaviour in coordinator layout
workaround for issue #26, it is based on
http://stackoverflow.com/questions/30923889/flinging-with-recyclerview-a
ppbarlayout

the solution on stack overflow did not work for me. the onNestedFling
is called when the fling is started, at that point the location of the
scroll offset is not really relevant, neither is the position in the
recycler view.
Therefor I opted to simulate a non-consumed fling in case when the user
starts flinging down